ebuild: Add support for building pinned cros-workon ebuilds
This change uses the CROS_WORKON_TREE information to generate
repo_repository rules and updates the ebuild targets to declare the
git_trees that point to those repositories. The git_trees are bind
mounted into the container, and cros-workon deals with extracting the
tarball into the ebuild's work directory.

This is dependent on tg/1753797.

BUG=b:269606353, b:267209587
TEST=BOARD=skyrim bazel run @portage//sys-boot/libpayload:0.0.1-r4296_debug -- --login=after
TEST=BOARD=skyrim bazel run @portage//chromeos-base/chromeos-config:0.0.2-r194

Change-Id: Ic117d941ac8d60e626df083d0bc72d96b4512132
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1753826
Reviewed-by: Ryo Hashimoto <hashimoto@google.com>
Reviewed-by: Shuhei Takahashi <nya@google.com>