ebuild: Generate the @chrome repositories
Since the only thing we need to fetch a chrome repository is the tag,
we can generate the repository rules for the chrome packages. This
removes the hard coded version we are currently using so that we can
build ToT without having to manually keep the version in sync.

One concern I have with adding the version number into the repository
name is that the .cache directory will end up with a bunch of stale
chrome versions. I'm not sure if there is a way to periodically clean
the cache. If this becomes a problem we can modify the code to generate
a single @chrome repository and error out if multiple versions are
required.

BUG=b:262460834
TEST=BOARD=arm64-generic bazel build @portage//chromeos-base/chromeos-chrome
and verify repository gets generated

Change-Id: I3751d70dc9717b5df2b8bde327b73ddd42f628a1
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1764439
Reviewed-by: Ryo Hashimoto <hashimoto@google.com>