Start reworking build_image rule
As we go through prototyping, the abstraction in mountsdk.bzl to provide
common logic for building packages and images got some skews. This patch
is a part of series of patches to fix them.

This patch simply removes the build_image rule temporarily to prepare
for refactoring. We will soon reintroduce it.

BUG=b:271493722
TEST=BOARD=arm64-generic bazel build @portage//sys-apps/attr

Change-Id: I6ab9f90b933bb9ea06d79dee57afc7ce46317284
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1767357
Reviewed-by: Ryo Hashimoto <hashimoto@google.com>
Reviewed-by: Raul Rangel <rrangel@google.com>