Start reworking mountsdk.bzl
As we go through prototyping, the abstraction in mountsdk.bzl to provide
common logic for building packages and images got some skews. This patch
is a part of series of patches to fix them.

This patch inlines mountsdk.bzl to various existing files to undo the
current abstraction. We will reintroduce mountsdk.bzl with a different
abstraction soon.

This is a pure refactoring change. It should not change actual behavior.

BUG=b:271493722
TEST=BOARD=arm64-generic bazel build @portage//sys-apps/attr

Change-Id: I112a23183b009b2f0aaf1e8d57f58f0013ebb49f
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1767360
Reviewed-by: Ryo Hashimoto <hashimoto@google.com>
Reviewed-by: Raul Rangel <rrangel@google.com>