Make BinaryPackageInfo hold references to BinaryPackageInfo
Currently, package_set rule abuses BinaryPackageInfo to represent a
set of packages, instead of a single package, by assigning None to the
`file` field. This is an ugly hack and does not work well in general.

Also, one thing I'd like to do next is to allow specifying package_set's
as well as individual packages in the build_image rule, e.g.

  build_image(
      name = "my_image",
      packages = [
          # Include base/dev/test packages.
          "@portage//virtual/target-os:package_set",
          "@portage//virtual/target-os-dev:package_set",
          "@portage//virtual/target-os-test:package_set",
	  # Also include the codelab package.
	  "@portage//chromeos-base/codelab",
      ],
      ...
  )

For this to work, `ebuild` rule and `package_set` rule should emit a
common provider.

For these reasons, I'll introduce a new provider BinaryPackageSetInfo
in the next CL as a way to represent a package set properly. However,
we cannot use Depset[Target] there to represent a package set because
the `ebuild` rule cannot get Target of the current target being
analyzed. Therefore, this patch changes BinaryPackageInfo to refer to
other packages with BinaryPackageInfo, not Target.

BUG=b:271493722
TEST=BOARD=arm64-generic bazel build @portage//sys-apps/acl

Change-Id: I4402cc5ed0a921dfbb5f0dd79197ab7b45aad8e7
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1770218
Reviewed-by: Ryo Hashimoto <hashimoto@google.com>
Reviewed-by: Raul Rangel <rrangel@google.com>