Replace executable=True with test=True
Convert the target's type to a test.
Can be run with `bazel run` with, but still unable to be run with `bazel
test` because of sandbox setup difference.

Also change how generate_test_runner.sh handles RUNFILE_DIR envvar to be
conpatible with how bazel runs targets with `test = True`.

BUG=b:269558613
TEST=BOARD=amd64-generic bazel run @portage//sys-apps/acl:test

Change-Id: I032421f6d101b47b880542922d92f797c7f806fa
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1776529
Reviewed-by: Shuhei Takahashi <nya@google.com>