sdk: Bump SDK version
ToT requires a newer SDK because we need the `rust-toolchain-version`
binary provided by the dev-lang/rust-host package.

This change adds a flag that can be used to toggle between the newer
SDK and the previous version.

I've also included a simple bash script that we can use to generate
part of the prebuilts/repositories.bazel file. This will need a lot more
love to actually generate a full SDK revbump, but it works for now.

BUG=b:271315440
TEST=BOARD=arm64-generic bazel run --//bazel/sdk:new-sdk @portage//dev-rust/assertions:debug -- --login=after

Change-Id: I6586e9f40c7cad153812fac620299c1be738d372
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1774791
Reviewed-by: Shuhei Takahashi <nya@google.com>