Fix meson_test.py hack
We originally had a wrapper that makes meson_test.py a no-op. In
https://team.git.corp.google.com/cros-build-tiger/third_party/eclass-overlay/+/cf2b7b5ca2fe086435be556c2685e7f8334305bc
I made it so the meson config dropped this wrapper so we didn't need to
override meson_test.py. Now that I'm rebasing, I don't want to hack
meson.eclass on ToT, so we need to restore the meson_test.py hack.

The hack can't be included in the extra_tarball that gets installed in
the base SDK because otherwise it gets deleted. makechroot::clean_layer
cleans up /mnt/host, so the meson_test.py gets removed from the SDK
layer.

This CL makes it so the hack is instead injected as a `src` tarball
to the ebuilds that need it. This will unfortunately mean all ebuilds
that use meson, will need to depend on platform2/common-mk and chromite
once we remove the hack :/

BUG=b:271315440, b:272275535
TEST=BOARD=arm64-generic bazel run  --//bazel/sdk:new-sdk @portage//dev-libs/libevdev:debug -- --login=after

Change-Id: Ica798f955e9e2b277c312bb0839ddc21b60e6293
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1774792
Reviewed-by: Shuhei Takahashi <nya@google.com>