sdk: Add b2 as a prebuilt to the SDK
b2 is required to build boost.

I think we need to add some sort of BDEPEND support. The hard coded list
of packages we install into the base SDK could be moved to the
individual packages that depend on the build tools. This would keep
our base SDK a bit smaller and prevent having to rebuild all the
packages if we need to add a new prebuilt packge.

BUG=b:271315440
TEST=BOARD=arm64-generic bazel run  --//bazel/sdk:new-sdk @portage//dev-libs/boost:debug -- --login=after

Change-Id: I94b0b9479976feebe746e0ee8b9e3037e4b1d97e
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1774793
Reviewed-by: Shuhei Takahashi <nya@google.com>