Install dependency packages to sysroot before building images
In the current build workflow with Portage, packages are all
installed to sysroot (/build/$BOARD) when we run build_image.
Some ebuilds assume this and can break otherwise. For example,
chromeos-fonts scans $SYSROOT/usr/bin/fonts to check for installed
fonts. I think this is rather a bug in the ebuild because they should
check for $ROOT instead, but in general it's more stable to replicate
what we do today in our ephemeral chroot.

This patch installs packages to sysroot before running build_image.

BUG=b:271929630
TEST=BOARD=amd64-generic bazel build //:chromiumos_base_image

Change-Id: I7c0c771e0f49069f4f896e3dd2388f4b864cb0e5
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1778637
Reviewed-by: Raul Rangel <rrangel@google.com>