alchemist: Add PackageAtom
`PackageAtom` should be used when parsing configuration files or
accepting user input. This is a sub-set of PackageDependencyAtom since
it doesn't support the blocking syntax, USe constraints or slot
operators. I used the existing PackageDependencyParser instead
of writing a new one since the syntax is similar. I think we might need
to fork it so we can support wildcard patterns correctly, but we can do
that later if we need it.

BUG=b:272366761
TEST=cargo test

Change-Id: I24a66e18bd20d8e087e8afb3b7dcee6956acf155
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1780970
Reviewed-by: Ryo Hashimoto <hashimoto@google.com>
Reviewed-by: Shuhei Takahashi <nya@google.com>