alchemist: Implement USE dependency matching
I changed the matches impl to return a Result. This is required because
we need to fail analysis if the USE flags aren't defined correctly in
the src or target USE maps. I also did some manual formatting on the
errors so we can keep errors constrained to a single line when we print
them with --verbose.

BUG=b:272366761
TEST=~/chromiumos$ BOARD=arm64-generic bazel query --//bazel/sdk:new-sdk @portage//virtual/target-os:all
* udev now gets correctly selected
TEST=~/cros-bazel$ BOARD=arm64-generic bazel query @portage//virtual/target-os:all
* mesa was replaced by minigbm in some cases
* haskell dropped a bunch of stuff in the package_set. Guessing it's
  using the binary package now.
* ipython got dropped because of unsatisfied USE constraints.
TEST=manually diff @portage directories

Change-Id: I76f24429b393f00836773763fe57606cba58d4b5
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1783130
Reviewed-by: Shuhei Takahashi <nya@google.com>