Reverse the representation order of filesystem layers
When overlayfs gets multiple lower directories, it treats the first
directory as the most "significant"; that is, when the same file name
is found in multiple lower directories, one from the earliest lower
directory takes precedence. This ordering is kind of intuitive for
overlayfs because we can interpret it as overlayfs looking up the given
lower directories one by one in the order.

We used to follow the ordering everywhere in our code, but we found that
the ordering is a bit un-intuitive and inefficient. The primary reason
is that our code constructs file system layers (not looking up) and thus
appending a newly created, most significant layer to the end of a layer
list is more intuitive.

This patch reverses the ordering of filesystem layers throughout our
code. Instead, run_in_container will concatenate lower directory paths
in the reversed order to construct an option string to pass to overlayfs.

BUG=None
TEST=BOARD=amd64-generic bazel build //:chromiumos_minimal_image

Change-Id: I6368900acf8e29a359e7de0b674a095e42635a1a
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1807529
Reviewed-by: Tomasz Tylenda <ttylenda@google.com>
Reviewed-by: Ryo Hashimoto <hashimoto@google.com>
Reviewed-by: Raul Rangel <rrangel@google.com>