ebuild: Print total number of dependencies that need to be installed
It's a little frustrating when it takes multiple minutes for
dependencies to install. By printing out the total number, it gives the
user an estimate at how expensive the task is. I think it will also help
encourage users to prune the dep graph if they see their package depends
on too m any packages.

This change does require flattening the depset, but I think it's a
worthwhile trade off.

> Installing 83 dependencies for cros-camera-libcamera_connector_test-9999.ebuild; 60s local
> Installing 172 dependencies for chromeos-chrome-114.0.5697.0_rc-r1.ebuild; 134s local
> Installing 86 dependencies for attestation-9999.ebuild; 5s local
> Installing 96 dependencies for shill-9999.ebuild; 4s local

BUG=b:277375685
TEST=BOARD=arm64-generic bazel build -k --config=prebuilts/arm64-generic  --//bazel/sdk:new-sdk @portage//media-libs/cros-camera-frame-annotator

Change-Id: Ib80012af45ba9b147d9848448c45d529fb06c081
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1819192
Reviewed-by: Shuhei Takahashi <nya@google.com>