Introduce durable trees
This patch introduces a new crate durabletree that works with durable
trees, a special directory format designed to preserve file metadata in
Bazel tree artifacts.

An arbitrary directory tree can be converted to a durable tree, and a
durable tree can be then converted to a set of directories that can be
mounted with overlayfs to reproduce the original directory. A durable
tree is safe as a Bazel tree artifact; that is, it does not contain
non-regular files (such as symlinks and character devices), and contains
metadata database to restore file metadata (permissions and xattrs for
now).

If a Bazel build action wants to output a directory to be mounted by
later actions, it can save the directory as a durable tree so that it is
reliably reproduced even if it is uploaded and downloaded from Bazel
remote caches.

BUG=b:277285193
TEST=bazel test :all

Change-Id: I9afc4af23574512ec2f79f15d083cf492f61aea9
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1829370
Reviewed-by: Ryo Hashimoto <hashimoto@google.com>