run_in_container: Run dumb-init as PID=1
unshare(CLONE_NEWPID) does NOT move the current process into a new PID
namespace, but it is its first child process created with fork/clone
that spawns into a new PID namespace as PID=1. In order for dumb-init
to be PID=1, run_in_container must run it as a subprocess.

BUG=None
TEST=BOARD=amd64-generic bazel build @portage//sys-apps/attr
TEST=grep NSpid /proc/${pid-of-dumb-init}/status

Change-Id: If1f06c729595985cbee27acba1867053966e02fd
Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/cros-bazel/+/1837590
Reviewed-by: Raul Rangel <rrangel@google.com>