alchemist: Generate an @portage symlink in workspace root
It's kind of a pain right now to find the generated @portage repo.
Especially with the migration to bzlmod. The @portage repo doesn't exist
in any of the generated `bazel-` directories. This change
creates a @portage symlink in the workspace root. This makes it easier
for people to browse the generated repo. As a nice side effect, it also
fixes tab completion!

The alternative to this CL is to do the following:
    $ ls bazel-out/../../../external/_main~portage~portage/

This will be difficult for people to discover and remember.

In order to prevent `bazel build //...` from unexpectedly building
this repository, I'm guarding it behind an environment flag. This way
people will be aware of this case.

BUG=b:264578578
TEST=BOARD=arm64-generic bazel query @portage//:all
TEST=BOARD=arm64-generic bazel query @portage//<tab><tab>

Change-Id: I3cd809e25d3dd527a6400a4785cb3e0a3ff8b37f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4595138
Reviewed-by: Matt Stark <msta@google.com>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Tim Bain <tbain@google.com>