Revert "Move processes::run to action_wrapper."
This reverts commit 82d84c56581f813d5c16fcb1509022c590496bdd.

processes::run needs to be used any time we spawn a child processes.
This ensures our processes tree correctly cleans up and we don't leave
any orphan processes still running that are adopted by init.

This change makes the :debug target function correctly on exit.

I changed the API to return an ExitStatus instead of Output. The Output
API was wrong since we were not reading from the output pipe. This means
we could have stalled the processes if the output buffer got full.

I also had to fix the mount_sdk test.

BUG=b:266919327
TEST=BOARD=amd64-host bazel build  --//bazel/sdk:new-sdk //bazel/sdk
TEST=BOARD=amd64-host bazel run  --//bazel/sdk:new-sdk @portage//dev-libs/glibc:debug

Change-Id: I1f0b5e379baccc30486b666a5ac3df0d509234e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4471187
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>