alchemist: Move packages into //internal/packages/<overlay>/<package>
Instead of encoding the overlay's mount path in the target path, we
instead add a reference to the overlay to each ebuild and use that to
look up the overlay's mount path. This decouples the location of the
ebuild's BUILD file and where it gets mounted inside the container. This
change helps clean up a lot of the path mangling we were doing.

This is subtle, but the inner ebuild path we pass into build_packages is
now an absolute path instead of a relative path. This kind of mirrors
what we do with the overlay layers, except the overlay
tarballs contain the absolute path encoded in the tarball. This also
means that `build_package` no longer cares about /mnt/host/source.

BUG=b:278728702
TEST=cargo test
TEST=BOARD=amd64-host bazel build  --//bazel/sdk:new-sdk //bazel/sdk:stage2

Change-Id: Ib2a3941276a22a73fcca7166057d5f954160ad7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4486802
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>