alchemist: Add a load_board function
In order to support host tools, we will also need to load the amd64-host
board. This change refactors the code a bit so we can easily load and
analyze a second board.

I created a TargetData struct that contains the config, loader,
resolver, etc that are all specific to a single board. This way we can
easily pass it around and know which board we are working on. I didn't
know what to call it honestly, so the name is pretty generic.

Since rust doesn't allow self-referential structs, I had to switch over
to using Arc. This allows us to clean up some extra .clones() and drop
some lifetime parameters.

BUG=b:278728702
TEST=cargo test
TEST=BOARD=arm64-generic bazel build  @portage//virtual/target-os:all

Change-Id: If2c51ba5444256df07eab3545c686820c7d349c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4492826
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>