alchemist: Move overlay_set from sdk to //internal/overlays
With the support for host tools, we now have two sets of overlays. This
change pulls them out of the SDK BUILD.bazel and into their own
BUILD.bazel file. I chose the name of the primary repository as the name
of the overlay_set because it's possible to build with:

    BOARD=amd64-host bazel build @portage//<target>

This will result in only a single overlay_set being generated.
Eventually when we add support for compiled profiles this all goes away.

I also cleaned up the board usage in the package template.

BUG=b:278728702
TEST=cargo test
TEST=BOARD=arm64-generic bazel build @portage//sys-libs/zlib
TEST=BOARD=arm64-generic bazel build @portage//internal/packages/stage1/target/bootstrap/portage-stable/sys-libs/zlib:1.2.13-r2

Change-Id: I8bb18697dd299f53bbe6df9634cdaec90ab8b351
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4510432
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>