ebuild: Move board arg from mountsdk to each CLI
In order to support host tools, we need to make the board argument
optional. This change moves the board argument to each CLI so that
we can migrate each CLI to support an optional board if necessary.
When the board is not specified, the operation will be performed on the
host.

The MountedSDK::new method now takes in an optional board argument. It
uses this to set the BOARD environment variable, and also sets the
correct ROOT. We should probably unset the $BOARD variable after we
consume it so it's not exposed to ebuilds. They should never be
referencing it.

BUG=b:278728702
TEST=cargo test
TEST=BOARD=arm64-generic bazel build @portage//sys-libs/zlib

Change-Id: I6298c1194cec28194130eebc3d72cbf3a917dedc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4514739
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>