run_in_container: Pass environment variables explicitly
Previously, to set environment variables for the program inside the
container, mountsdk used to set environment variables for
run_in_container because the program inside the container inherited
environment variables of run_in_container. This approach came with
a few trickiness because run_in_container itself depends on several
environment variables, such as TMPDIR and RUNFILES_DIR, while we
don't want to expose them to the program inside the container.

This patch introduces a new field to RunInContainerConfig that
explicitly carries environment variables for the program inside the
container. This way, we can set different environment variables for
the program inside the container and run_in_container, which makes
things easier and cleaner.

BUG=b:284387940
TEST=BOARD=amd64-generic bazel build //bazel/images:chromiumos_base_image

Change-Id: I0a6201b79f9a849d3fef63dfe385b15a921b6c6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4574342
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>