run_in_container: Speed up by leaving a sentinel process
When under heavy load, the kernel takes multiple seconds to destroy a
network namespace. Since the parent run_in_container process is the last
process in the network namespace, it may take some time between
run_in_container exits and its parent process is notified for the
process exit and wastes time.

This is a hacky patch to work around the problem by leaving a "sentinel"
process in the network namespace. This will shorten the time to install
0 build-time dependencies to ~1 second even under high load.

BUG=b:284922501
TEST=BOARD=amd64-generic bazel build @portage//sys-apps/attr

Change-Id: I4d0730d549934ea4dd557647f12670f688f51e7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4583835
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>