ebuild: Make sdk_update work with python 3.8
ToT has added python 3.8, so we need to patch that version of portage.
Patching python is pretty messy... some things use lib or lib64 and it's
not really consistent. We should work on landing the patches upstream
so we can drop this workaround, and all of sdk_update.

BUG=b:278728702
TEST=BOARD=arm64-generic bazel build -k @portage//internal/sdk/stage2

Change-Id: Iae9ce420ee6f63c3595804b46f4627016b2626cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4618442
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>