container: Support allow_network_access and privileged
The privilege option in the new API has difference from the old API:
it does not run sudo, and instead requires the current process to
already have privilege. This is because running sudo there interferes
with the usage of enter_mount_namespace that enters a user namespace.

BUG=b:286150000
TEST=bazel test :all

Change-Id: If2c24bc1176b9aef27d1fa19aa0ec91567b8765b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4603431
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>