container: Rename the staging directory
Currently, the container crate and its users (such as build_package)
use the same directory (/mnt/host/bazel-build) to stage some files.
It works with the old API (mountsdk.rs) but not with the new API
(container.rs) due to the difference of how to stage those files.
In any case, there is no strong reason to use the same directory,
so let us just rename it.

BUG=b:286150000
TEST=bazel test :all

Change-Id: Iad08fd6c9713222910045a3fbb8d4cb2352e2cf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4603435
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>