Add extract_package program
extract_package takes a binary package file as an input, and outputs its
installed image as a durable tree.

An installed image is a file system layer that we can mount with
overlayfs to simulate installing a package. It includes files contained
in the tarball part of the binary package, as well as VDB directory
structure that would be created by Portage on install.

BUG=b:280673549
TEST=bazel test //bazel/portage/bin/extract_package:all

Change-Id: I24817193851a45b37bc94969e3e1f88fa8d6c973
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4768100
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>