alchemist: Introduce PackageType to allow specifying package types
We can now specify if we are generating host packages or target
packages. We can use this information to generate BDEPEND/IDEPEND for
the ebuild targets. This is implemented by adding a new _host_deps SDK
layer that purely installs host tools. Since we also know how many deps
we are installing we can skip generating an SDK layer if we don't have
any deps to install. In order to avoid circular dependencies, we need
to strip out the dependencies for packages that are provided by the
host SDK layer. To aid in debugging, we also print out the provided
packages in a comment.

BUG=b:278728702
TEST=./bazel/tools/run_tests.sh

Change-Id: I438237b6c2e0682a04a44f83ea4414b52b78cb6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4608510
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>