alchemist: Generate stage 2 host SDK
The stage2/host SDK differs from the stage2 SDK in that the profile is
different. When building packages we want to use the overlay-amd64:base
profile. This is because the profile needs to inherit from the
toolchains:base profile so we can pick up the correct cross-compiler
config.

When building target packages, we don't care about the toolchain config.
If we were to use the overlay-amd64:base profile it would require that
we include the overlay-amd64 and toolchains overlays because portage
parses the host's profile so it can compute the CBUILD variable. So
the stage2 SDK uses the default chromiumos:default/linux/amd64/10.0/sdk
profile. Since this profile is always present we can avoid including
any additional overlays.

Once we have compiled profiles we should be able to generate a "host"
profile that only contains the `CHOST` variable so portage can compute
the `CBUILD` variable when doing a cross-root compile.

BUG=b:278728702
TEST=BOARD=arm64-generic bazel build @portage//internal/sdk/stage2/host
TEST=BOARD=arm64-generic bazel build @portage//internal/packages/stage2/host/toolchains/cross-aarch64-cros-linux-gnu/gcc:10.2.0-r32

Change-Id: I4d51787cddc1387480649201fd42388702911330
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4618443
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>