alchemist: Generate stage 2 target SDK
This is the SDK that will be used to build the packages for the target
board.

With this change we can finally start building target packages using
the stage2 SDK! This means we now properly support BDEPENDs. There
are still a lot of packages that don't declare their BDEPENDs, and a few
ebuild changes that need to land before this is fully functional.

BUG=b:278728702
TEST=BOARD=arm64-generic bazel build @portage//internal/sdk/stage2/target/board
TEST=BOARD=arm64-generic bazel build -k @portage//internal/packages/stage2/target/board/chromiumos/virtual/target-os:1_package_set
TEST=stabilize-15429.B$ BOARD=amd64-generic bazel build -k @portage//sys-libs/zlib
TEST=stabilize-15429.B (chroot)$ BOARD=amd64-generic ~/go/bin/bazelisk build //bazel/images:chromiumos_minimal_image

Change-Id: Ie0f2219ac11bc976eed7fa8d3cc7a6a4bd8ddd1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4619682
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>