ebuild: Manually install cross glibc in stage1 target SDK
The legacy CrOS SDK populates package.provided with glibc. When we are
executing inside the chroot we use the chroot's portage config instead
of the fakeroot config we generate. This means we need to match the
fakeroot config as close as possible. In order to fully close out
b/266979761, we will need to modify the legacy setup_board script to
stop treating glibc as a special case. For now we need to restore the
behavior where we use the cross toolchain's glibc.

BUG=b:286202533
TEST=ToT$ BOARD=amd64-generic bazel build //bazel/images:chromiumos_minimal_image
     ToT build image is broken. Not related to this chain.
TEST=stabilize-15429.B$ BOARD=amd64-generic bazel build //bazel/images:chromiumos_minimal_image
TEST=(chroot) stabilize-15429.B$ BOARD=amd64-generic ~/go/bin/bazelisk build //bazel/images:chromiumos_minimal_image

Change-Id: I73dc95d9cc95927cf842df1453f6e6773df94f2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4630038
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>