action_wrapper: Handle errors correctly
Formerly action_wrapper failed without printing anything to stderr
when it encounters an error in the middle (e.g. when the specified
program is not found) and --log is specified.

This patch fixes the issue by making sure to print logs even when
it encounters errors.

BUG=b:274042271
TEST=bazel test :all

Change-Id: I5aac16a57dfae6ffec084819fb56fe3ec6edb6cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4666904
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>