build_sdk: Add a workaround for /usr/lib64 symlink
We still have some packages that rely on /usr/lib being a symlink to
lib64, so we need this to maintain comparability with the CrOS SDK.
I can't figure out which package is responsible for creating it, so for
now I'm adding a work around. This allows rust-host to build
successfully.

BUG=b:278728702
TEST=BOARD=arm64-generic bazel build @portage//internal/packages/stage2/host/chromiumos/dev-lang/rust-host:1.69.0-r1

Change-Id: I2116b5acffab472eb16c7634226b6290c4b75ceb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4652392
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>