alchemist: Add a workaround for dev-libs/nss
When cross-compiling dev-libs/nss, it requires the build host to have
dev-libs/nss installed. This constraint only applies when
cross-compiling, when doing a native build, we don't need dev-libs/nss
pre installed. We can't add dev-libs/nss to the BDEPEND line because
that would cause a circular dependency when building for ROOT=/.

This change works around the problem by introducing a BDEPEND that
is only active when we are cross-compiling. Portage doesn't have a way
of expressing this type of dependency.

BUG=b:278728702
TEST=BOARD=amd64-generic bazel build @portage//internal/packages/stage2/target/board/chromiumos/dev-libs/nss:3.68.2-r3

Link: https://bugs.gentoo.org/759127
Change-Id: Id0091fe96e23150b2edf3e903d30ece5e8f58d2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4655206
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>