alchemist: Track all the sourced files from make.conf
We want to have a complete picture of all the files that we used to
compute the config. We weren't tracking the additional files that were
sourced in make.conf or make.default files. In order to track the
multiple files that make up the make.conf, we need to keep a vec of
sources in the ConfigNode. I also got rid of the ConfigNode::new
constructor since it was only used in once place.

BUG=b:284339358
TEST=cargo test

Change-Id: I20016f5ec4fa7236347ead9e57a650b0b2840dbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4658614
Commit-Queue: Ryo Hashimoto <hashimoto@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>