alchemist: Generate IDEPENDs
When installing the dev-libs/nss package into the board's sysroot, we
need to ensure the IDEPENDs are installed into the host_deps SDK layer.
This change makes it so we take the IDEPENDs into account when
generating the list of packages we install into the `_host_deps` SDK
layer. I chose to compute this in alchemist instead of bazel since the
calculation would be complex to perform in bazel and it would make the
`sdk_install_deps` API ugly since we would need to also list all the
package's DEPENDs so we could extract the IDEPENDs.

BUG=b:278728702
TEST=BOARD=amd64-generic bazel build @portage//internal/packages/stage2/target/board/chromiumos/chromeos-base/libchrome:9999_deps

Change-Id: I10849bdb75fb57ac5fbfbe1571ae26225c2b85f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4669898
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>