alchemist: Use DEPEND as BDEPEND when EAPI < 7
I was trying to avoid doing this, but we have quite a large number of
packages that haven't been migrated to EAPI 7. When a package is older
than EAPI 7, we need to use the DEPEND as the BDEPEND. This
unfortunately pulls in a lot of extra host packages that are not really
BDEPENDS. In order to not pull in libchrome and even chromeos-chrome
as BDEPENDS, I've created a curated allow list so we can control which
dependencies we pull in. Once we migrate all packages to EAPI 7+ we
can drop this hack.

With this change we can now cross-compile a lot of older packages.

BUG=b:278728702
TEST=cargo run -- --board arm64-generic generate-repo --output-dir ~/tmp/stage-depend --output-repos-json ~/tmp/repos.json

Change-Id: I961a98e220ed11647a663fccf9de06717891ee1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4673476
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>