Abort Alchemy package build if any ebuild failed
When at least one of ebuilds for some package failed to be analyzed
by Alchemy, point the package's alias target to the failure(s).

This is a change of behavior in the situation when a package has
multiple ebuilds and only some of them failed to be analyzed. Before
this commit, Alchemy would silently fall back to choosing the best
version among successfully parsed ebuilds. After this commit, if a
package needs to be built and at least one of its ebuilds failed to
parse, the resulting build will fail as well. This is a stricter,
but also safer behavior.

Example of the message:

  Analyzing: target @_main~portage~portage//acct-user/chronos:chronos (1 packages loaded, 0 targets configured)
  ERROR: <...>/acct-user/chronos/BUILD.bazel:68:12: in always_fail rule @_main~portage~portage//acct-user/chronos:failure:
  Traceback (most recent call last):
          File "<...>/always_fail.bzl", line 8, column 9, in _always_fail_impl
                  fail(ctx.attr.message)
  Error in fail:
  --
  Some of ebuilds failed to analyze! Aborting the build to prevent possibly incorrect ebuild selection.
  target: @@_main~portage~portage/acct-user/chronos
  List of failures below:

  ** chronos-0.0.2.ebuild **
  ebuild failed to evaluate <...>/chronos-0.0.2.ebuild: exit status: 1
  stdout:
  stderr: FATAL: HEHE

  ** chronos-999.ebuild **
  ebuild failed to evaluate <...>/chronos-999.ebuild: exit status: 1
  stdout:
  stderr: FATAL: FOO

  --

BUG=b:290769888
TEST=create acct-user/chronos/chronos-9999.ebuild with "die FOO" statement, run "BOARD=amd64-generic bazel build @portage//acct-user/chronos", verify it prints the error instead of building chronos-0.0.2
TEST="BOARD=amd64-generic bazel build @portage//virtual/target-os:package_set" succeeds

Change-Id: Ibebb227f94adbf63d778bc8c981d3814b7ddba23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4675555
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>