alchemist: Allow get_extra_dependencies to specify BDEPEND for EAPI <7
We can't currently can't add any missing BDEPENDs using
get_extra_dependencies. This is because we are only evaluating the
DEPEND as BDEPEND for EAPI < 7. This change makes it so we evaluate the
BDEPEND and DEPEND as BDEPEND and merge the two lists. This does end up
allowing an EAPI < 7 ebuild to specify BDEPEND, but I didn't want to
make the code any more complicated.

BUG=b:278728702
TEST=cargo run -- --board arm64-generic generate-repo --output-dir ~/tmp/stage-depend --output-repos-json ~/tmp/repos.json

Change-Id: I4f263b16d1c8f09104e86c164f4cefb0c101b0e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4695118
Tested-by: Raul Rangel <rrangel@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>